requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
joblib==1.5.0
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.16.0
slack_sdk==3.26.2
threadpoolctl==3.6.0
tzdata==2023.4
urllib3==2.2.0
